编译告警清理
git-svn-id: http://192.168.20.11/sequoiadb/trunk@30079 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/dmsStorageDataCapped.hpp
@@ -398,7 +398,7 @@ namespace engine
       SDB_ASSERT( mbStatInfo, "mbStatInfo should not be NULL" ) ;
 
       return ( mbStatInfo->_totalDataLen + newSize ) >
-               _options[context->mbID()]->_maxSize ;
+               (UINT64)_options[context->mbID()]->_maxSize ;
    }
 
    OSS_INLINE BOOLEAN _dmsStorageDataCapped::_numExceedLimit( dmsMBContext *context,
@@ -409,7 +409,7 @@ namespace engine
 
       return ( _options[context->mbID()]->_maxRecNum > 0 &&
                ( ( mbStatInfo->_totalRecords + newNum ) >
-                 _options[context->mbID()]->_maxRecNum ) ) ;
+                 (UINT64)_options[context->mbID()]->_maxRecNum ) ) ;
    }
 
    OSS_INLINE BOOLEAN _dmsStorageDataCapped::_overwriteOnExceed( dmsMBContext *context )
